+++ scripts/test/Ticket/TicketHistory.t
@@ -27,6 +27,12 @@ $Kernel::OM->ObjectParamAdd(
 );
 my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
+# Turn on the ticket type feature.
+$Kernel::OM->Get('Kernel::Config')->Set(
+    Key   => 'Ticket::Type',
+    Value => 1,
+);
+
 $Kernel::OM->Get('Kernel::System::Cache')->CleanUp();
 
 my @Tests = (
@@ -92,6 +98,19 @@ my @Tests = (
                         HistoryType => 'NewTicket',
                         Type        => 'Unclassified',
                     },
+
+                    # Bug 12702 - TicketHistoryGet() initial ticket type update
+                    {
+                        CreateBy    => 1,
+                        HistoryType => 'TypeUpdate',
+                        Queue       => 'Raw',
+                        OwnerID     => 1,
+                        PriorityID  => 3,
+                        State       => 'new',
+                        Type        => 'Unclassified',
+                        TypeID      => '1',
+                    },
+
                     {
                         CreateBy    => 1,
                         HistoryType => 'CustomerUpdate',
@@ -311,6 +330,7 @@ for my $Test (@Tests) {
 
             my %LookForHistoryTypes = (
                 NewTicket      => 1,
+                TypeUpdate     => 1,
                 OwnerUpdate    => 1,
                 CustomerUpdate => 1,
             );